+++ scripts/jest/createCacheKeyFunction.js
@@ -20,19 +20,9 @@ function buildCacheKey(files, base) {
   );
 }
 
-const transformRoot = path.join(__dirname, '..');
-const cacheKeyFiles = [
-  __filename,
-  path.join(transformRoot, 'babel-6', 'default-options.js'),
-  path.join(transformRoot, 'babel-6', 'dev-expression.js'),
-  path.join(transformRoot, 'babel-6', 'inline-requires.js'),
-  path.join(transformRoot, 'babel-6', 'rewrite-modules.js'),
-];
-
-const cacheKeyBase = buildCacheKey(cacheKeyFiles, '');
-
 module.exports = files => {
-  const cacheKey = buildCacheKey(files, cacheKeyBase);
+  const presetVersion = require('../package').dependencies['babel-preset-fbjs'];
+  const cacheKey = buildCacheKey(files, presetVersion);
   return (src, file, configString) => crypto.createHash('md5')
     .update(cacheKey)
     .update(src + file + configString)