+++ scripts/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changed
 - [lint] Upgraded config to match latest internal config. Requires ESLint >= v2.0.0 (may be using newer rules).
+- [babel] Moved all plugins into new `babel-preset-fbjs` package
 
 
 ## [0.6.0] - 2016-04-04